feat(viewer): split igneous-md into server and viewer components
