+++ Cargo.toml
@@ -1,3 +1,15 @@
 [workspace]
 members = ["crates/*"]
 resolver = "2"
+
+[workspace.package]
+version = "0.2.0"
+edition = "2021"
+
+[workspace.dependencies]
+
+igneous-md-viewer = { version = "0.1.0", path = "crates/igneous-md-viewer" }
+igneous-md = { version = "0.2.0", path = "crates/igneous-md" }
+
+[workspace.metadata.crane]
+name = "igneous-md-workspace"
