+++ rustfmt.toml
@@ -1,0 +1,1 @@
+merge_imports = true
